# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iqrah Backend API",
    "description": "REST API for the Iqrah mobile app",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/releases/{id}/publish": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Publishes a release after validation.",
        "operationId": "publish_release",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Release ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Release published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishReleaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Release validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/v1/admin/releases/{id}/validate": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Validates release publish constraints.",
        "operationId": "validate_release",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Release ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateReleaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid/disabled admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "admin_api_key": []
          },
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/releases/latest": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Returns the latest published release and required artifacts.",
        "operationId": "get_latest_release",
        "responses": {
          "200": {
            "description": "Latest published release",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LatestReleaseResponse"
                }
              }
            }
          },
          "404": {
            "description": "No published release found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/releases/{id}/manifest": {
      "get": {
        "tags": [
          "releases"
        ],
        "summary": "Returns the full artifact manifest for one published release.",
        "operationId": "get_release_manifest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Release ID (UUID)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Release manifest",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseManifestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid release id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Release not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "API error response format",
        "required": [
          "error"
        ],
        "properties": {
          "details": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "example": "[\"name: validation failed\"]"
          },
          "error": {
            "type": "string",
            "example": "Validation error: field is required"
          }
        }
      },
      "ArtifactRole": {
        "type": "string",
        "description": "Artifact semantic role within a dataset release.",
        "enum": [
          "core_content_db",
          "knowledge_graph",
          "morphology",
          "translation_catalog",
          "audio_catalog",
          "optional_pack"
        ],
        "example": "core_content_db"
      },
...
      "DatasetRelease": {
        "type": "object",
        "description": "Dataset release entity.",
        "required": [
          "id",
          "version",
          "status",
          "created_by",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T16:00:00Z"
          },
          "created_by": {
            "type": "string"
          },
          "id": {
            "$ref": "#/components/schemas/ReleaseId"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "example": "2026-02-20T17:00:00Z"
          },
          "status": {
            "$ref": "#/components/schemas/DatasetReleaseStatus"
          },
          "version": {
            "type": "string",
            "example": "2026.02.20.1"
          }
        }
      },
...
      "DatasetReleaseStatus": {
        "type": "string",
        "description": "Dataset release lifecycle status.",
        "enum": [
          "draft",
          "published",
          "deprecated"
        ],
        "example": "draft"
      },
...
      "LatestReleaseResponse": {
        "type": "object",
        "description": "Latest published release with required artifacts.",
        "required": [
          "release",
          "required_artifacts"
        ],
        "properties": {
          "release": {
            "$ref": "#/components/schemas/DatasetRelease"
          },
          "required_artifacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseArtifactManifestEntry"
            }
          }
        }
      },
...
      "PackId": {
        "type": "string",
        "description": "Stable pack identifier.",
        "example": "translation.en"
      },
...
      "PublishReleaseResponse": {
        "type": "object",
        "description": "Publish response for a release.",
        "required": [
          "published",
          "release"
        ],
        "properties": {
          "published": {
            "type": "boolean"
          },
          "release": {
            "$ref": "#/components/schemas/DatasetRelease"
          }
        }
      },
...
      "ReleaseArtifactManifestEntry": {
        "type": "object",
        "description": "Artifact manifest entry used by clients.",
        "required": [
          "package_id",
          "required",
          "artifact_role",
          "version",
          "sha256",
          "file_size_bytes",
          "download_url"
        ],
        "properties": {
          "artifact_role": {
            "$ref": "#/components/schemas/ArtifactRole"
          },
          "download_url": {
            "type": "string",
            "example": "http://localhost:8080/v1/packs/translation.en/download"
          },
          "file_size_bytes": {
            "type": "integer",
            "format": "int64",
            "example": 1024
          },
          "package_id": {
            "$ref": "#/components/schemas/PackId"
          },
          "required": {
            "type": "boolean"
          },
          "sha256": {
            "type": "string",
            "example": "ba7816bf8f01cfea414140de5dae2223b00361a396177a9cb410ff61f20015ad"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          }
        }
      },
      "ReleaseId": {
        "type": "string",
        "description": "Stable dataset release identifier.",
        "example": "9f9f88ce-f30b-4f7f-9ea9-2de9e53ca6f1"
      },
      "ReleaseManifestResponse": {
        "type": "object",
        "description": "Full manifest for one release.",
        "required": [
          "release",
          "artifacts"
        ],
        "properties": {
          "artifacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseArtifactManifestEntry"
            }
          },
          "release": {
            "$ref": "#/components/schemas/DatasetRelease"
          }
        }
      },
      "ReleaseValidationIssue": {
        "type": "object",
        "description": "One validation issue from release checks.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "missing_required_roles"
          },
          "message": {
            "type": "string",
            "example": "Required artifact roles are missing: knowledge_graph"
          }
        }
      },
      "ReleaseValidationReport": {
        "type": "object",
        "description": "Result of release validation.",
        "required": [
          "valid",
          "failures",
          "warnings"
        ],
        "properties": {
          "failures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseValidationIssue"
            }
          },
          "valid": {
            "type": "boolean"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReleaseValidationIssue"
            }
          }
        }
      },
...
      "ValidateReleaseResponse": {
        "type": "object",
        "description": "Validation response for a release.",
        "required": [
          "report"
        ],
        "properties": {
          "report": {
            "$ref": "#/components/schemas/ReleaseValidationReport"
          }
        }
      }
    },
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}